# public/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: BE_Laravel_8
  contact: {}
  version: '1.0'
servers:
  - url: https://be-laravel.herokuapp.com/api
    variables: {}
paths:
...
  /destroy/2/delete:
    delete:
      tags:
        - Misc
      summary: destroy
      operationId: destroy
      parameters:
        - name: Accept
          in: header
          description: ''
          required: true
          style: simple
          schema:
            type: string
            example: application/json
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
        - bearer: []
...
components:
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...